apps/api/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
python-dotenv>=1.0
aiosqlite>=0.20
